OpenID is now optional
New setting: $Config['OpenID'] (boolean), enables OpenID if true.
Default value: (unset)

A 403 forbidden is returned if we try to achieve an OpenID transaction
and OpenID is disabled.

Follow-up commit should update the OpenID code to comply with the latest
pear code, or use Composer to offer an alternative package.

[Commit done in the 138b and 136a TEC Namur lines.]