Removed hardcoded domain from a dev helper script
The objects viewer requires the path to the API to work. As it's intended
for any development server and not only zed51, domain has been removed but
path / kept, as we can safely consider every dev installation puts Zed in
the root path when the goal is to test the position of the objects and not
the queer installation scenarios.