[Geo] Fixes incomplete scene content file issue
Each scene content file must contain location information in header.

This piece of information allows to build an index
mapping files and locations.

Such header usually takes the form of an HTML comment:
    <!--
    Global location: B00001002
    Local location: (0, 0, 0)
    -->

If not, the following warning occured:
Warning: strpos(): Offset not contained in string
in .../includes/geo/sceneindex.php on line 174

We now throw an exception if the global location isn't set.

Fixes T97.