[Session] Don't encode IP addresses as hexadecimal strings
Summary:
Previously, Keruald stored IP addresses as hexadecimal strings.

The benefit of this operation is dubious, and this doesn't work
for IPv6. So, instead to encode IPv6 and increase db field size,
this change stores directly IP adresses in canonical format.

This change is a temporary hotfix, as session code has already
been refactored for Keruald, and need to be integratd to Zed.

A SQL schema update is needed:
`ALTER TABLE sessions CHANGE IP IP varchar(45) NOT NULL DEFAULT '';`

Test Plan:
Browse https://zed51.dereckson.be with IPv6