[Smarty] Don't add current directory in scene rendering
To render a scene, a Smarty template located in the scene directory
is rendered. To do so, with the idea such directory is a relative path
to the current web site, we added the web site root as template dir.

Nowadays, Smarty seems to be glad to consider a tempalte path defined
by a relative path according the current working directory out of the box.

This code isn't needed anymore.

Furthermore, this change fixes the following issue introduced by breaking
compatibility changes in Smarty 3.1:

```
[15-Apr-2018 21:11:46 UTC] PHP Warning:  rtrim() expects parameter 1 to be string,
array given in /var/51-wwwroot/zed/vendor/smarty/smarty/libs/Smarty.class.php on line 1078
```