Fix PHP_SELF nginx issue for get current url fragments
In the previous nginx and Apache configuration, PHP_SELF matched
the index.php entry point. As such, PHP_SELF was used to determine
if the current URL is the entry point.

With the current nginx configuration, PHP_SELF and REQUEST_URI
match. As such, this tweak can't be used anymore.