Extract geo methods into hypership/geo library
Summary:
The code for cartesian/spherical/cylindrical coordinates has been
extracted and moved to a new library, hypership/geo.

This change removes the library code for the Zed codebase
and supersedes it by calls to the library classes.

A Smarty modified plugin 'sector' avoids direct call to Octocube
class into a template.

The legacy GeoOctocube class contains a method specific to Zed
to generate a MySQL search clause for a sector and a specific height.
This has been moved to a new OctocubeBuilder engine.

Test Plan: Unit tests provided for new plugin.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://code.zed.dereckson.be/D16