Enable interactive debugger
Add "?debug=true" to the URL can control redirect workflow
for the XHR entry point, allowing to check result before being
redirected.

This is not always convenient, for example to test content upload,
URL is generated by JS script.

This change adds to the smartline a "debug [on|off]" command to store
in the session whether the debugger is enabled.