Ensure localisation code is PHP 8 ready
Wnile the intersection code is candidate for Keruald Omnitools,
Zed will still own most of the localisation code, to be moved
in a specific engine.

The find_lang function has now the responsibility to always
return a language, or throw an exception when files are missing.

Add parameters types, return type, etc.