Refactor models in Models namespace
The models in includes/objects, includes/content and includes/geo
are now moved to a Zed\Models namespace.

As part of an ongoing effort to decouple the dependencies,
each model receives a DatabaseEngine engine in constructor
and for static helper methods.

The controllers and engines are responsibles to inject
the database dependency.